requirements.txt
aiohttp==3.10.5
pyjwt==2.9.0
datetime==5.1